ulong -> decimal

+++ FibonacciOnConsole/IFibonacci.cs
@@ -3,6 +3,6 @@ namespace FibonacciOnConsole
 {
     public interface IFibonacciStrategy
     {
-        ulong Calc(ulong i);
+        decimal Calc(decimal i);
     }
 }
